+++ CHANGES.md
@@ -9,7 +9,15 @@
 
 ## Changelogs
 
-### v0.2.3 - **Current**
+### v0.2.4 - **Current**
+
+- Graphics:
+  - New `Vertex Array` resource
+- Geometry:
+  - New namespace `Coli::Geometry` for geometry features
+  - New `Vertex` struct and its traits `VertexTraits`
+
+### v0.2.3
 
 - Graphics:
   - New `Program` resource